+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -37,7 +37,6 @@ public class AddressBookParser {
      * @throws ParseException if the user input does not conform the expected format
      */
     public Command parseCommand(String userInput) throws ParseException {
-        System.out.println(userInput);
         final Matcher matcher = BASIC_COMMAND_FORMAT.matcher(userInput.trim());
         if (!matcher.matches()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, HelpCommand.MESSAGE_USAGE));
@@ -46,7 +45,6 @@ public class AddressBookParser {
         final String commandWord = matcher.group("commandWord");
 
         final String arguments = matcher.group("arguments");
-        System.out.println(arguments);
 
         switch (commandWord) {
