+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -1,13 +1,10 @@
 package seedu.address.ui;
 
-import java.awt.Desktop;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
 import java.util.List;
 import java.util.logging.Logger;
+import javax.swing.JFileChooser;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.MenuItem;
@@ -16,20 +13,14 @@ import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.Logic;
-import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.SummariseCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.AddCommandParser;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.excel.ExcelToJSONConverter;
 import seedu.address.model.excel.ImportFileParser;
-import java.util.concurrent.TimeUnit;
-import javax.swing.*;
 
 /**
  * The Main Window. Provides the basic application layout containing
@@ -38,6 +29,7 @@ import javax.swing.*;
 public class MainWindow extends UiPart<Stage> {
 
     private static final String FXML = "MainWindow.fxml";
+    private static HelpWindow helpWindow;
 
     private final Logger logger = LogsCenter.getLogger(getClass());
 
@@ -48,10 +40,8 @@ public class MainWindow extends UiPart<Stage> {
     private PersonListPanel personListPanel;
     private ResultDisplay resultDisplay;
     private EmailWindow emailWindow;
-    private static HelpWindow helpWindow;
     private ImportWindow importWindow;
 
-    public static File excelFile;
 
     @FXML
     private PieChartWindow pieChartWindow;
@@ -173,10 +163,10 @@ public class MainWindow extends UiPart<Stage> {
     }
 
     /**
-     * Opens file manager for import
+     * Opens file manager for import window
      */
     @FXML
-    public void handleImport() throws ParseException, CommandException, InterruptedException {
+    public void handleImport () throws ParseException, CommandException {
         JFileChooser fileChooser = new JFileChooser();
         int returnValue = fileChooser.showOpenDialog(null);
         if (returnValue == JFileChooser.APPROVE_OPTION) {
@@ -184,8 +174,7 @@ public class MainWindow extends UiPart<Stage> {
             assert (selectedFile != null);
             ImportFileParser converter = new ImportFileParser();
             List<String> res = converter.JsonToPerson(selectedFile);
-            System.out.println(res);
-            for(int i = 0; i < res.size(); i++) {
+            for (int i = 0; i < res.size(); i++) {
                 executeCommand(res.get(i));
             }
         }
