+++ src/test/java/seedu/address/logic/commands/EditPersonDescriptorTest.java
@@ -1,4 +1,3 @@
-
 package seedu.address.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -63,5 +62,3 @@ public class EditPersonDescriptorTest {
 
     }
 }
-
-
